WebSockets/tests/submissions/Opera/interfaces/WebSocket/events/016.html

<!doctype html>
<title>WebSockets: addEventListener</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':80/ws/echo');
  var count = 0;
  function checkCount(c, e) {
    count++;
    debug(e.type + ', '+c);
    assert_equals(count, c);
  }
  // no spec requires this order for event listeners but the web does
  ws.addEventListener('open', function(e) {
    checkCount(1, e); 
     ws.send('Goodbye');
  }, false);
  ws.onopen = function(e) {checkCount(2, e) };
  ws.addEventListener('open', function(e) {checkCount(3, e); }, false);
  
  ws.addEventListener('message', function(e) {checkCount(4, e); }, false);
  ws.onmessage = function(e) {checkCount(5, e) };
  ws.addEventListener('message', function(e) {checkCount(6, e); }, false);
  
  ws.addEventListener('close', function(e) {checkCount(7, e); }, false);
  ws.onclose = function(e) {checkCount(8, e) };
  ws.addEventListener('close', function(e) {
    checkCount(9, e);
    end();
  }, false);
  
});
</script>
